requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-bootstrap4==1.1.1
psycopg2-binary==2.8.4
pytz==2019.3
soupsieve==2.0
sqlparse==0.3.0
django-crispy-forms==1.9.0
